refactor: consolidate activity domain core (#176)
