+++ models.py
@@ -1,38 +1,10 @@
-from dataclasses import asdict, dataclass, field
-from typing import Any, Dict, List, Optional, Tuple
+"""Compatibility wrapper for the activity domain model.
 
+The canonical activity model now lives in :mod:`qfit.activities.domain.models`.
+Keep this module as a stable import surface while the package structure is
+refactored incrementally.
+"""
 
-@dataclass
-class Activity:
-    source: str
-    source_activity_id: str
-    external_id: Optional[str] = None
-    name: Optional[str] = None
-    activity_type: Optional[str] = None
-    sport_type: Optional[str] = None
-    start_date: Optional[str] = None
-    start_date_local: Optional[str] = None
-    timezone: Optional[str] = None
-    distance_m: Optional[float] = None
-    moving_time_s: Optional[int] = None
-    elapsed_time_s: Optional[int] = None
-    total_elevation_gain_m: Optional[float] = None
-    average_speed_mps: Optional[float] = None
-    max_speed_mps: Optional[float] = None
-    average_heartrate: Optional[float] = None
-    max_heartrate: Optional[float] = None
-    average_watts: Optional[float] = None
-    kilojoules: Optional[float] = None
-    calories: Optional[float] = None
-    suffer_score: Optional[float] = None
-    start_lat: Optional[float] = None
-    start_lon: Optional[float] = None
-    end_lat: Optional[float] = None
-    end_lon: Optional[float] = None
-    summary_polyline: Optional[str] = None
-    geometry_source: Optional[str] = None
-    geometry_points: List[Tuple[float, float]] = field(default_factory=list)
-    details_json: Dict[str, Any] = field(default_factory=dict)
+from .activities.domain.models import Activity
 
-    def to_record(self) -> Dict[str, Any]:
-        return asdict(self)
+__all__ = ["Activity"]
